Update source api java

+++ gender/src/main/java/databases/mysql/repositories/ILanguageRepository.java
@@ -2,8 +2,10 @@ package databases.mysql.repositories;
 
 import databases.entities.Language;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ILanguageRepository extends JpaRepository<Language, Long> {}
+public interface ILanguageRepository extends JpaRepository<Language, BigInteger> {}
